+++ packages/Bio/src/utils/monomer-lib/library-file-manager/file-validator.ts
@@ -45,6 +45,7 @@ export class MonomerLibFileValidator {
 
   private validateJsonContent(jsonContent: any[], fileName: string): boolean {
     let isValid = true;
+    const existingMonomerSymbols = new Set<string>();
     for (const monomer of jsonContent) {
       const name = monomer[REQ.SYMBOL] ?? monomer[REQ.ID] ?? monomer[REQ.NAME] ?? NA_CODE;
       isValid = this.validateMonomerSchema(monomer);
@@ -59,6 +60,13 @@ export class MonomerLibFileValidator {
         );
         break;
       }
+      const key = `${(monomer[REQ.POLYMER_TYPE] ?? '')}-${name}`;
+      if (existingMonomerSymbols.has(key)) {
+        console.warn(`Bio: Monomer Library File Validator file ${fileName}, monomer '${name}' is duplicated.`,
+          'Please, verify that the monomer library file does not contain duplicated monomer symbols.'
+        );
+      }
+      existingMonomerSymbols.add(key);
     }
     return isValid;
   }