+++ packages/Bio/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Add getHelm for Cyclized notation provider
 * Add detector for Dimerized notation provider
+* Add Monomer library manager view with duplicate preferences
+* Add Monomer manager for editing/adding/removing monomers
 
 ### Bug fixes